+++ fugitive-darkness-plugin/fugitive-darkness-plugin-vertx-git-provider/src/main/java/com/zer0s2m/fugitivedarkness/plugin/vertx/git/provider/GitRepoProviderConverter.java
@@ -2,6 +2,7 @@ package com.zer0s2m.fugitivedarkness.plugin.vertx.git.provider;
 
 import com.zer0s2m.fugitivedarkness.plugin.vertx.git.provider.impl.GitRepoProviderConverterImpl;
 import com.zer0s2m.fugitivedarkness.provider.ContainerInfoRepo;
+import com.zer0s2m.fugitivedarkness.provider.GitProviderNotFoundException;
 import com.zer0s2m.fugitivedarkness.provider.GitRepoProviderType;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
@@ -21,8 +22,10 @@ public interface GitRepoProviderConverter {
      * @param providerType Provider type.
      * @param payload      Response from the repository provider.
      * @return Repository information.
+     * @throws GitProviderNotFoundException Git provider not found.
      */
-    ContainerInfoRepo convert(final GitRepoProviderType providerType, JsonObject payload);
+    ContainerInfoRepo convert(final GitRepoProviderType providerType, JsonObject payload)
+            throws GitProviderNotFoundException;
 
     /**
      * Convert the response from the repository provider into a familiar object for the system.
@@ -35,7 +38,13 @@ public interface GitRepoProviderConverter {
         List<ContainerInfoRepo> result = new ArrayList<>();
         payload
                 .stream()
-                .forEach((item) -> result.add(convert(providerType, (JsonObject) item)));
+                .forEach((item) -> {
+                    try {
+                        result.add(convert(providerType, (JsonObject) item));
+                    } catch (GitProviderNotFoundException e) {
+                        throw new RuntimeException(e);
+                    }
+                });
         return result;
     }
 
@@ -43,9 +52,11 @@ public interface GitRepoProviderConverter {
      * Set provider host.
      *
      * @param providerType Provider type.
-     * @param host Host.
+     * @param host         Host.
+     * @throws GitProviderNotFoundException Git provider not found.
      */
-    void setHost(final GitRepoProviderType providerType, String host);
+    void setHost(final GitRepoProviderType providerType, String host)
+            throws GitProviderNotFoundException;
 
     static GitRepoProviderConverter create() {
         return new GitRepoProviderConverterImpl();