+++ fugitive-darkness-provider/src/main/java/com/zer0s2m/fugitivedarkness/provider/impl/GitRepoImpl.java
@@ -2,6 +2,7 @@ package com.zer0s2m.fugitivedarkness.provider.impl;
 
 import com.zer0s2m.fugitivedarkness.common.Environment;
 import com.zer0s2m.fugitivedarkness.provider.*;
+import org.eclipse.jgit.api.CloneCommand;
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.errors.GitAPIException;
 import org.eclipse.jgit.lib.ObjectId;
@@ -17,6 +18,10 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicInteger;
 
 /**
@@ -29,28 +34,45 @@ public class GitRepoImpl implements GitRepo {
     /**
      * Clone a git repository from a remote host into a set path environment variable {@link Environment#ROOT_PATH_REPO}.
      *
-     * @param URI Remote git repository host.
+     * @param command Command to clone a repository.
+     * @param URI     Remote git repository host. Must not be {@literal null}.
      * @return Git repository information.
-     * @throws GitAPIException The exception is caused by the internal functionality of managing git repositories.
+     * @throws GitAPIException he exception is caused by the internal functionality of managing git repositories.
      */
     @Override
-    public ContainerInfoRepo gClone(String URI) throws GitAPIException {
+    public ContainerInfoRepo gCloneStart(
+            CloneCommand command, String URI) throws GitAPIException {
         final ContainerInfoRepo infoRepo = gGetInfo(URI);
 
-        logger.info("Start cloning the repository: " + infoRepo.source());
+        logger.info("Start cloning the repository [{}:{}]: {}",
+                infoRepo.group(), infoRepo.project(), infoRepo.source());
 
-        Git.cloneRepository()
+        command
+                .call()
+                .close();
+
+        logger.info("Finish cloning the repository [{}:{}]: {}",
+                infoRepo.group(), infoRepo.project(), infoRepo.source());
+
+        return infoRepo;
+    }
+
+    /**
+     * Create a command to clone a repository from a remote host.
+     *
+     * @param URI Remote git repository host. Must not be {@literal null}.
+     * @return Command.
+     */
+    @Override
+    public CloneCommand gCloneCreate(String URI) {
+        final ContainerInfoRepo infoRepo = gGetInfo(URI);
+
+        return Git.cloneRepository()
                 .setURI(URI)
                 .setDirectory(infoRepo.source().toFile())
                 .setCloneAllBranches(true)
                 .setCloneSubmodules(true)
-                .setNoCheckout(true)
-                .call()
-                .close();
-
-        logger.info("Finish cloning the repository: " + infoRepo.source());
-
-        return infoRepo;
+                .setNoCheckout(true);
     }
 
     /**
@@ -155,28 +177,11 @@ public class GitRepoImpl implements GitRepo {
         final List<ContainerInfoSearchGitRepo> searchFileGitRepos = new ArrayList<>();
         filterSearch.getSources()
                 .forEach(source -> {
+                    logger.info("Start the search [{}]", source);
+
                     final ContainerGitRepoMeta gitRepo = filterSearch.getGitMeta(source);
                     try {
                         final SearchEngineGitGrep commandGrep = searchEngineGitGrep(filterSearch, source, gitRepo);
-
-                        if (filterSearch.getPatternForIncludeFile() != null) {
-                            commandGrep.setPatternForIncludeFile(filterSearch.getPatternForIncludeFile());
-                        }
-
-                        if (filterSearch.getPatternForExcludeFile() != null) {
-                            commandGrep.setPatternForExcludeFile(filterSearch.getPatternForExcludeFile());
-                        }
-
-                        commandGrep.setMaxCount(filterSearch.getMaxCount());
-                        commandGrep.setMaxDepth(filterSearch.getMaxDepth());
-
-                        if (filterSearch.getContext() == -1 || filterSearch.getContext() == 0) {
-                            commandGrep.setContextBefore(filterSearch.getContextBefore());
-                            commandGrep.setContextAfter(filterSearch.getContextAfter());
-                        } else {
-                            commandGrep.setContext(filterSearch.getContext());
-                        }
-
                         final List<ContainerInfoSearchFileGitRepo> searchResult = commandGrep.callGrep();
 
                         searchFileGitRepos.add(new ContainerInfoSearchGitRepo(
@@ -187,10 +192,52 @@ public class GitRepoImpl implements GitRepo {
                                 commandGrep.getExtensionFilesGrep(),
                                 searchResult
                         ));
+
+                        logger.info("End of search    [{}]", source);
                     } catch (IOException | SearchEngineGitException e) {
                         throw new RuntimeException(e);
                     }
                 });
+        return searchFileGitRepos;
+    }
+
+    /**
+     * Search for matches in files in git repositories by pattern. Git grep command.
+     * <p>Uses a search engine {@link SearchEngineGitGrep}.</p>
+     *
+     * @param filterSearch Filter for searching git repositories.
+     * @return Search result in git repository.
+     */
+    @Override
+    public List<ContainerInfoSearchGitRepo> searchByGrepVirtualThreads(GitRepoFilterSearch filterSearch) {
+        final List<ContainerInfoSearchGitRepo> searchFileGitRepos = new ArrayList<>();
+
+        try (final ExecutorService executor = Executors.newVirtualThreadPerTaskExecutor()) {
+            final List<GitRepoSearchCallable> gitRepoSearchCallables = new ArrayList<>();
+            filterSearch.getSources().forEach((source) -> {
+                final GitRepoFilterSearch newFilterSearch = GitRepoFilterSearch.clone(filterSearch);
+                newFilterSearch.clearGitMeta();
+                newFilterSearch.clearSources();
+                newFilterSearch.addGitRepo(source);
+                newFilterSearch.addGitMeta(source, filterSearch.getGitMeta(source));
+
+                gitRepoSearchCallables.add(new GitRepoSearchCallable(newFilterSearch));
+            });
+
+            final List<Future<List<ContainerInfoSearchGitRepo>>> futures =
+                    executor.invokeAll(gitRepoSearchCallables);
+
+            futures.forEach((future) -> {
+                try {
+                    searchFileGitRepos.addAll(future.get());
+                } catch (InterruptedException | ExecutionException e) {
+                    throw new RuntimeException(e);
+                }
+            });
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        }
+
         return searchFileGitRepos;
     }
 
@@ -201,12 +248,16 @@ public class GitRepoImpl implements GitRepo {
      * @param source       Source path of the git repository.
      * @param gitRepo      More information about the git repository.
      * @return Assembled search engine.
-     * @throws IOException If an IO error occurred.
+     * @throws IOException                         If an IO error occurred.
+     * @throws SearchEngineGitSetMaxCountException Exception for setting the maximum number of matches in one file.
+     * @throws SearchEngineGitSetMaxDepthException Exception for setting maximum search depth.
+     * @throws SearchEngineGitSetContextException  Exception for setting preview code after and before matches.
      */
     private static SearchEngineGitGrep searchEngineGitGrep(
             GitRepoFilterSearch filterSearch,
             Path source,
-            ContainerGitRepoMeta gitRepo) throws IOException {
+            ContainerGitRepoMeta gitRepo) throws IOException, SearchEngineGitSetMaxCountException,
+            SearchEngineGitSetMaxDepthException, SearchEngineGitSetContextException {
         final SearchEngineGitGrep commandGrep = new SearchEngineGitGrepImpl(
                 filterSearch.getPattern(),
                 source,
@@ -218,6 +269,25 @@ public class GitRepoImpl implements GitRepo {
         if (filterSearch.getExcludeExtensionFile() != null) {
             commandGrep.setExcludeExtensionFilesForSearchGrep(filterSearch.getExcludeExtensionFile());
         }
+
+        if (filterSearch.getPatternForIncludeFile() != null) {
+            commandGrep.setPatternForIncludeFile(filterSearch.getPatternForIncludeFile());
+        }
+
+        if (filterSearch.getPatternForExcludeFile() != null) {
+            commandGrep.setPatternForExcludeFile(filterSearch.getPatternForExcludeFile());
+        }
+
+        commandGrep.setMaxCount(filterSearch.getMaxCount());
+        commandGrep.setMaxDepth(filterSearch.getMaxDepth());
+
+        if (filterSearch.getContext() == -1 || filterSearch.getContext() == 0) {
+            commandGrep.setContextBefore(filterSearch.getContextBefore());
+            commandGrep.setContextAfter(filterSearch.getContextAfter());
+        } else {
+            commandGrep.setContext(filterSearch.getContext());
+        }
+
         return commandGrep;
     }