+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## [0.0.4] 2024-02-02
+
+### Added
+
+- Basic exceptions for git providers.
+- Search engine:
+    - Clone repositories from a remote host (From git providers).
+    - Use search splitting into virtual threads.
+
+### Fixed
+
+- Fix pack empty repository for search.
+- Fix search exception - MissingObjectException.
+
 ## [0.0.3] 2024-01-28
 
 ### Added